fix(driver): PT_UID is of type uint32_t

+++ driver/ppm_fillers.c
@@ -9347,7 +9347,7 @@ int f_sys_setuid_x(struct event_filler_arguments *args) {
 	int64_t retval = 0;
 	int64_t res = 0;
 	unsigned long val = 0;
-	int32_t uid;
+	uint32_t uid;
 
 	/* Parameter 1: res (type: PT_ERRNO) */
 	retval = (int64_t)syscall_get_return_value(current, args->regs);
@@ -9356,8 +9356,8 @@ int f_sys_setuid_x(struct event_filler_arguments *args) {
 
 	/* Parameter 2: uid (type: PT_UID) */
 	syscall_get_arguments_deprecated(args, 0, 1, &val);
-	uid = (int32_t)val;
-	res = val_to_ring(args, (int64_t)uid, 0, false, 0);
+	uid = (uint32_t)val;
+	res = val_to_ring(args, uid, 0, false, 0);
 	CHECK_RES(res);
 
 	return add_sentinel(args);
